dropPmidVersionColumn now works for sqlite databases.

+++ R/utilities.R
@@ -116,7 +116,23 @@ getPkgVersion = function(pkgName = 'pmparser') {
 dropPmidVersionColumn = function(tableSuffix, con) {
   emptyTables = getEmptyTables(tableSuffix)
   idx = !grepl('^(pmid_status|xml_processed)', names(emptyTables))
-  for (tableName in names(emptyTables)[idx]) {
-    q = sprintf('alter table %s drop column version', tableName)
-    x = DBI::dbExecute(con, q)}
+
+  if (attr(class(con), 'package') == 'RSQLite') { # thanks, sqlite
+    for (tableName in names(emptyTables)[idx]) {
+      cols = setdiff(DBI::dbListFields(con, tableName), 'version')
+      tableTmp = paste_(tableName, 'tmp')
+
+      q = sprintf('create table %s as select %s from %s',
+                  tableTmp, paste(cols, collapse = ', '), tableName)
+      x = DBI::dbExecute(con, q)
+      DBI::dbRemoveTable(con, tableName)
+
+      q = sprintf('alter table %s rename to %s', tableTmp, tableName)
+      x = DBI::dbExecute(con, q)}
+
+  } else {
+    for (tableName in names(emptyTables)[idx]) {
+      q = sprintf('alter table %s drop column version', tableName)
+      x = DBI::dbExecute(con, q)}}
+
   invisible()}
